Add endpoint for updating wishlist items
- Implement PUT /wishlists/{id}/items/{item_id}
- Add validation for item ownership
- Include comprehensive error handling
- Add unit tests for happy and sad paths
