+++ tests/test_routes.py
@@ -441,8 +441,6 @@ class TestSadPaths(TestCase):
         response = self.client.post(BASE_URL, json=wishlist_data)
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
 
-    # Todo: Add your test cases here...
-
     def test_update_item_not_found(self):
         """It should return 404 if the item does not exist"""
         wishlist = WishlistFactory()
@@ -491,3 +489,25 @@ class TestSadPaths(TestCase):
         wishlist.create()
         resp = self.client.put(f"{BASE_URL}/{wishlist.id}")
         self.assertEqual(resp.status_code, status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)
+
+    def test_update_nonexistent_wishlist(self):
+        """It should return 404 when updating a non-existent wishlist"""
+        response = self.client.put(
+            f"{BASE_URL}/99999",
+            json={"name": "Test", "customer_id": "test"},
+            content_type="application/json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+    def test_add_item_no_content_type(self):
+        """It should return 415 when adding item without content type"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        response = self.client.post(f"{BASE_URL}/{wishlist.id}/items")
+        self.assertEqual(response.status_code, status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)
+
+    def test_get_wishlist_item_wrong_wishlist(self):
+        """It should return 404 when getting item from wrong wishlist"""
+        # check wishlist_id match
+        response = self.client.get(f"{BASE_URL}/99999/items/99999")
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
