+++ tests/test_wishlist_item.py
@@ -17,6 +17,7 @@ DATABASE_URI = os.getenv(
 ######################################################################
 #        W I S H L I S T  I T E M   M O D E L   T E S T   C A S E S
 ######################################################################
+# pylint: disable=too-many-public-methods
 class TestWishlistItem(TestCase):
     """WishlistItem Model Test Cases"""
 
@@ -102,7 +103,6 @@ class TestWishlistItem(TestCase):
         item = WishlistItemFactory(wishlist=wishlist)
         wishlist.wishlist_items.append(item)
         wishlist.create()
-        # Assert that it was assigned an id and shows up in the database
         self.assertIsNotNone(wishlist.id)
         wishlists = Wishlist.all()
         self.assertEqual(len(wishlists), 1)
@@ -233,7 +233,7 @@ class TestWishlistItem(TestCase):
         original_delete = db.session.delete
 
         def raise_exc(obj):
-            raise Exception("DB error")
+            raise ValueError("DB error")
 
         db.session.delete = raise_exc
         try:
@@ -249,7 +249,7 @@ class TestWishlistItem(TestCase):
         original_commit = db.session.commit
 
         def raise_exc():
-            raise Exception("DB error")
+            raise ValueError("DB error")
 
         db.session.commit = raise_exc
         try:
@@ -340,3 +340,170 @@ class TestWishlistItem(TestCase):
         item = WishlistItem()
         with self.assertRaises(DataValidationError):
             item.deserialize(["not", "a", "dict"])
+
+    def test_wishlist_item_create_db_exception(self):
+        """It should raise DataValidationError if db.session.add fails during create"""
+        item = WishlistItemFactory()
+        item.id = None
+        original_add = db.session.add
+
+        def raise_exc(obj):
+            raise ValueError("DB error")
+
+        db.session.add = raise_exc
+        try:
+            with self.assertRaises(DataValidationError):
+                item.create()
+        finally:
+            db.session.add = original_add
+
+    # Moved from test_wishlist.py, related to items
+    def test_add_wishlist_items(self):
+        """It should create a wishlist and assert that it exists"""
+        fake_wishlist = WishlistFactory()
+        # pylint: disable=unexpected-keyword-arg
+        wishlist = Wishlist(
+            name=fake_wishlist.name,
+            id=fake_wishlist.id,
+            is_public=fake_wishlist.is_public,
+            customer_id=fake_wishlist.customer_id,
+            description=fake_wishlist.description,
+            created_at=fake_wishlist.created_at,
+            updated_at=fake_wishlist.updated_at,
+        )
+        self.assertIsNotNone(wishlist)
+        self.assertEqual(wishlist.id, fake_wishlist.id)
+        self.assertEqual(wishlist.name, fake_wishlist.name)
+        self.assertEqual(wishlist.customer_id, fake_wishlist.customer_id)
+        self.assertEqual(wishlist.description, fake_wishlist.description)
+        self.assertEqual(wishlist.created_at, fake_wishlist.created_at)
+        self.assertEqual(wishlist.updated_at, fake_wishlist.updated_at)
+        self.assertEqual(wishlist.wishlist_items, fake_wishlist.wishlist_items)
+
+    def test_serialize_a_wishlist(self):
+        """It should serialize a wishlist with items"""
+        wishlist = WishlistFactory()
+        item = WishlistItemFactory(wishlist=wishlist)
+        wishlist.wishlist_items.append(item)
+        wishlist.create()
+        serial = wishlist.serialize()
+        self.assertEqual(serial["id"], wishlist.id)
+        self.assertEqual(serial["name"], wishlist.name)
+        self.assertEqual(serial["customer_id"], wishlist.customer_id)
+        self.assertEqual(serial["is_public"], wishlist.is_public)
+        self.assertEqual(len(serial["wishlist_items"]), 1)
+        self.assertEqual(serial["wishlist_items"][0]["product_name"], item.product_name)
+
+    def test_deserialize_a_wishlist(self):
+        """It should deserialize a wishlist from a dict with items"""
+        wishlist = WishlistFactory()
+        item = WishlistItemFactory(wishlist=wishlist)
+        wishlist_dict = wishlist.serialize()
+        wishlist_dict["wishlist_items"] = [item.serialize()]
+        new_wishlist = Wishlist()
+        new_wishlist.deserialize(wishlist_dict)
+        self.assertEqual(new_wishlist.name, wishlist.name)
+        self.assertEqual(new_wishlist.customer_id, wishlist.customer_id)
+        self.assertEqual(len(new_wishlist.wishlist_items), 1)
+        self.assertEqual(new_wishlist.wishlist_items[0].product_name, item.product_name)
+        self.assertEqual(new_wishlist.wishlist_items[0].product_id, item.product_id)
+        self.assertEqual(new_wishlist.wishlist_items[0].quantity, item.quantity)
+        self.assertEqual(
+            new_wishlist.wishlist_items[0].product_description, item.product_description
+        )
+        self.assertEqual(
+            float(new_wishlist.wishlist_items[0].product_price),
+            float(item.product_price),
+        )
+
+    def test_delete_a_wishlist(self):
+        """It should delete a wishlist from the database"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        wishlist_id = wishlist.id
+        wishlist.delete()
+        self.assertIsNone(Wishlist.find(wishlist_id))
+
+    def test_find_by_name(self):
+        """It should find wishlists by name"""
+        wishlist = WishlistFactory(name="SpecialName")
+        wishlist.create()
+        found = Wishlist.find_by_name("SpecialName")
+        self.assertGreaterEqual(len(found), 1)
+        self.assertEqual(found[0].name, "SpecialName")
+
+    def test_find_for_user(self):
+        """It should find wishlists for a specific user"""
+        wishlist = WishlistFactory(customer_id="user123")
+        wishlist.create()
+        found = Wishlist.find_for_user("user123")
+        self.assertGreaterEqual(len(found), 1)
+        self.assertEqual(found[0].customer_id, "user123")
+
+    def test_deserialize_wishlist_key_error(self):
+        """It should raise DataValidationError for missing required fields"""
+        wishlist = Wishlist()
+        with self.assertRaises(DataValidationError):
+            wishlist.deserialize({})
+
+    def test_deserialize_wishlist_type_error(self):
+        """It should raise DataValidationError for wrong type"""
+        wishlist = Wishlist()
+        with self.assertRaises(DataValidationError):
+            wishlist.deserialize([])
+
+    def test_deserialize_wishlist_items_type_error(self):
+        """It should raise DataValidationError if wishlist_items is not a list"""
+        wishlist = Wishlist()
+        data = {"name": "Test", "customer_id": "c1", "wishlist_items": "not_a_list"}
+        with self.assertRaises(DataValidationError):
+            wishlist.deserialize(data)
+
+    def test_wishlist_create_db_exception(self):
+        """It should raise DataValidationError if db.session.add or commit fails during create"""
+        wishlist = WishlistFactory()
+        wishlist.id = None
+        # Patch db.session.add to raise Exception
+        original_add = db.session.add
+
+        def raise_exc(obj):
+            raise ValueError("DB error")
+
+        db.session.add = raise_exc
+        try:
+            with self.assertRaises(DataValidationError):
+                wishlist.create()
+        finally:
+            db.session.add = original_add
+
+    def test_wishlist_update_db_exception(self):
+        """It should raise DataValidationError if db.session.commit fails during update"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        original_commit = db.session.commit
+
+        def raise_exc():
+            raise ValueError("DB error")
+
+        db.session.commit = raise_exc
+        try:
+            with self.assertRaises(DataValidationError):
+                wishlist.update()
+        finally:
+            db.session.commit = original_commit
+
+    def test_wishlist_delete_db_exception(self):
+        """It should raise DataValidationError if db.session.delete or commit fails during delete"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        original_delete = db.session.delete
+
+        def raise_exc(obj):
+            raise ValueError("DB error")
+
+        db.session.delete = raise_exc
+        try:
+            with self.assertRaises(DataValidationError):
+                wishlist.delete()
+        finally:
+            db.session.delete = original_delete
